+++ doc/versionHistory.rst
@@ -6,6 +6,22 @@
 Version History
 ##################
 
+.. _lsst.ts.wep-12.0.0:
+
+-------------
+12.0.0
+-------------
+
+* Change pandas.DataFrame outputs to Astropy Tables.
+
+.. _lsst.ts.wep-11.5.2:
+
+-------------
+11.5.2
+-------------
+
+* Added a ComCamSim production pipeline for testing purposes.
+
 .. _lsst.ts.wep-11.5.1:
 
 -------------